daemon: split fqdn specific functions into fqdn.go
[ upstream commit 38aa328283f3d128bcf225a057984ae8026a2495 ]

This is a simple move, with no funcitonal change.
